When a user types 1.5 servings it is now accepted and shown in the input field

+++ src/epicr.h
@@ -82,7 +82,7 @@ namespace epicr
 
     typedef struct servings_s
     {
-        int count;
+        double count;
         std::string descriptor;
     } servings;
